Extract accordion-form controller.
