+++ app/components/accordion_step_component.html.erb
@@ -1,4 +1,4 @@
-<%= tag.div id: "#{id}-accordion", class: %w[accordion-item my-4 shadow-sm].append("accordion-step-#{step_index}").concat(@classes), data: data.merge('patron-request-target': 'accordion') do %>
+<%= tag.div id: "#{id}-accordion", class: %w[accordion-item my-4 shadow-sm].append("accordion-step-#{step_index}").concat(@classes), data: data.merge('accordion-form-target': 'accordion') do %>
   <div class="accordion-header d-flex">
     <button class="accordion-button" type="button" data-bs-toggle="collapse" data-bs-target="#<%= id %>" aria-expanded="<%= expanded? ? 'true' : 'false' %>" aria-controls="<%= id %>" disabled>
       <div class="step-number d-flex justify-content-center align-items-center rounded-circle border p-2 me-2 text-white"><%= step_index %></div>
@@ -6,13 +6,13 @@
         <%= title %>
       </h2>
     </button>
-    <button class="btn btn-link edit-button align-items-center" data-action="click->patron-request#editForm analytics#send" data-analytics-category-param="Accordion" data-analytics-action-param="Edit" aria-controls="<%= id %>">
+    <button class="btn btn-link edit-button align-items-center" data-action="click->accordion-form#editStep analytics#send" data-analytics-category-param="Accordion" data-analytics-action-param="Edit" aria-controls="<%= id %>">
       <i class="bi bi-pencil me-1"></i> Edit
     </button>
   </div>
 
   <% if body? %>
-    <div id="<%= id %>" data-action="input->patron-request#enableNextButton"  class="accordion-collapse collapse <%= 'show' if expanded? %>" data-bs-parent="#<%= form_id %>">
+    <div id="<%= id %>" data-action="input->accordion-form#enableNextButton"  class="accordion-collapse collapse <%= 'show' if expanded? %>" data-bs-parent="#<%= form_id %>">
       <div class="accordion-body">
         <%= body %>
 
@@ -25,7 +25,7 @@
           </div>
         <% else %>
           <div class="mt-3 d-flex justify-content-end gap-1">
-            <button class="btn btn-primary" type="button" data-action="click->patron-request#nextStep analytics#send" data-analytics-category-param="Accordion" data-analytics-action-param="Continue" data-next-step-button="true">Continue</button>
+            <button class="btn btn-primary" type="button" data-action="click->accordion-form#nextStep analytics#send" data-analytics-category-param="Accordion" data-analytics-action-param="Continue" data-next-step-button="true">Continue</button>
           </div>
         <% end %>
       </div>
