+++ app/views/patron_requests/new_redesign.html.erb
@@ -73,7 +73,7 @@
 <% end %>
 
 <%= form_for(@patron_request,
-            data: { controller: 'patron-request' },
+            data: { controller: 'patron-request accordion-form' },
             html: { class: 'col-lg-8 accordion' },
             url: patron_requests_path,
             method: 'POST') do |f| %>
@@ -143,17 +143,17 @@
       <% if f.object.selectable_items.length >= 10 %>
         <div>
           <label for="item-selector-search" class="my-2">Search item list</label>
-          <input id="item-selector-search" type="search" incremental data-action="input->itemselector#filter analytics#send" data-analytics-category-param="Item Selector" data-analytics-action-param="Filter" class="form-control w-75" />
+          <input id="item-selector-search" type="search" incremental data-action="input->itemselector#filter analytics#send" data-analytics-category-param="Item Selector" data-analytics-action-param="Filter" class="form-control w-75">
         </div>
       <% end %>
       <div class="item-table my-4">
         <table class="table mb-0 border" data-controller="table" data-table-field-value="Callnumber" data-table-asc-value="true">
           <thead>
             <tr>
-              <th data-table-target="header" aria-sort="ascending" class="text-start" data-table-field-param="Callnumber" data-action='click->table#sort'>
+              <th data-table-target="header" aria-sort="ascending" class="text-start" data-table-field-param="Callnumber" data-action="click->table#sort">
                 <button type="button" class="btn btn-link pe-0 text-decoration-none text-black" data-action="analytics#send" data-analytics-category-param="Item Selector" data-analytics-action-param="Sort/Call number">Call number</button>
               </th>
-              <th data-table-target="header" class="text-end" data-table-field-param="Status" data-action='click->table#sort'>
+              <th data-table-target="header" class="text-end" data-table-field-param="Status" data-action="click->table#sort">
                 <button type="button" class="btn btn-link pe-0 text-decoration-none text-black" data-action="analytics#send" data-analytics-category-param="Item Selector" data-analytics-action-param="Sort/Status">Status</button>
               </th>
             </tr>
@@ -215,7 +215,7 @@
 
   <!-- pickup request (multiple items) -->
   <% if can?(:request_pickup, f.object) %>
-    <%= render AccordionStepComponent.new(request: f.object, id: 'pickup', classes: [('d-none' if scan_pickup)], form_id: f.id, step_index: (scan_or_pickup_step ||= step_enum.next), data: { 'patronrequest-forRequestType': 'pickup', action: 'itemselector:changed@document->patron-request#enableAnyNextButtons' }, submit: true) do |c| %>
+    <%= render AccordionStepComponent.new(request: f.object, id: 'pickup', classes: [('d-none' if scan_pickup)], form_id: f.id, step_index: (scan_or_pickup_step ||= step_enum.next), data: { 'patronrequest-forRequestType': 'pickup', action: 'itemselector:changed@document->accordion-form#enableAnyNextButtons' }, submit: true) do |c| %>
       <% c.with_title.with_content('Pickup request') %>
       <% c.with_body do %>
         <%= render 'pickup_destination', f: %>
@@ -361,4 +361,3 @@
     <% end %>
   <% end %>
 <% end %>
-
